feature(logger-orchestrator): Separated prompt component from error-boundary.
