+++ public/components/common/error-boundary/error-boundary.tsx
@@ -14,7 +14,6 @@
 
 import React, { Component } from 'react';
 import loglevel from 'loglevel';
-import { EuiEmptyPrompt } from '@elastic/eui';
 import { UI_LOGGER_LEVELS } from '../../../../common/constants';
 import {
   UI_ERROR_SEVERITIES,
@@ -23,6 +22,7 @@ import {
   UILogLevel,
 } from '../../../react-services/error-orchestrator/types';
 import { ErrorOrchestratorClass } from '../../../react-services';
+import { ErrorComponentPrompt } from '../error-boundary-prompt/error-boundary-prompt';
 
 export default class ErrorBoundary extends Component {
   constructor(props) {
@@ -37,7 +37,7 @@ export default class ErrorBoundary extends Component {
     const { errorTitle, style, errorInfo }: Readonly<any> = this.state;
 
     if (errorInfo) {
-      return <HandleError errorTitle={errorTitle} errorInfo={errorInfo} style={style} />;
+      return <ErrorComponentPrompt errorTitle={errorTitle} errorInfo={errorInfo} style={style} />;
     }
     return this.props.children;
   }
@@ -54,7 +54,7 @@ const catchFunc = (errorTitle, errorInfo, ctx) => {
       context: ctx.context,
       level: UI_LOGGER_LEVELS.WARNING as UILogLevel,
       severity: UI_ERROR_SEVERITIES.UI as UIErrorSeverity,
-      display: true,
+      display: false,
       store: true,
       error: {
         error: errorTitle.name,
@@ -69,37 +69,3 @@ const catchFunc = (errorTitle, errorInfo, ctx) => {
     loglevel.error('Logger failed: ', error);
   }
 };
-
-const ErrorComponent = (props: { errorTitle: any; errorInfo: any; style: any }) => {
-  const styles = {
-    error: {
-      borderTop: '1px solid #777',
-      borderBottom: '1px solid #777',
-      padding: '12px',
-    },
-  };
-
-  return (
-    <div style={props.style || styles.error}>
-      <details style={{ whiteSpace: 'pre-wrap' }}>
-        {props.errorTitle && props.errorTitle.toString()}
-        <br />
-        {props.errorInfo.componentStack}
-      </details>
-    </div>
-  );
-};
-
-const HandleError = (props: { errorTitle: any; errorInfo: any; style: any }) => (
-  <EuiEmptyPrompt
-    iconType="faceSad"
-    title={<h2>Something went wrong.</h2>}
-    body={
-      <ErrorComponent
-        errorTitle={props.errorTitle}
-        errorInfo={props.errorInfo}
-        style={props.style}
-      />
-    }
-  />
-);
